requirements.txt
python-dotenv
gradio
pyperclip
chromadb
langchain
langchain_community
langchain_core
sentence-transformers
crawl4ai
pymupdf
bs4
googlesearch-python
aiohttp
gtts
SpeechRecognition
PyAudio
scipy
soundfile